+++ R/zzz.R
@@ -7,7 +7,8 @@
 #' @importFrom assertthat not_empty assert_that
 #' @importFrom R6 R6Class
 #' @importFrom crayon green blue
-#' @import shiny spsUtil
+#' @importFrom utils installed.packages
+#' @import shiny
 #'
 NULL